Move nuki logic to separate task
